// spadata/spa-data/ngx-data-requests/src/shared/interfaces/detail-request.ts
export interface DetailRequest {
    petitionNumber: string;
    petitionDate: string;
    startDate: string;
    sucursal: string;
    format: string;
    petitionType: string;
    policy: string;
    subAgentNumber: string;
    receipt: string;
    result: string;
    name: string;
    text: string;
    email: string;
    origin: string;
    typeSelectionDate: string;
    status: string;
    userId: string;
    endDate: string;
    mediador: string;
    product: string;
    extractionType: string;
    code: string;
    subAgentPresence: string;
    downloadDate: string;
    visible: string;
    active: string;
}